Backend/backend/requirements.txt
defusedxml==0.4.1
Django==1.10.3
django-allauth==0.28.0
django-cors-headers==1.3.1
django-rest-auth==0.8.2
djangorestframework==3.5.3
oauthlib==2.0.0
python3-openid==3.0.10
requests==2.12.0
requests-oauthlib==0.7.0
six==1.10.0
wheel==0.24.0
